+++ HISTORY.md
@@ -1,6 +1,11 @@
+## 1.0.3 (2012-03-23)
+
+* handle OBJECT_LIBRARY targets introduced in CMake 2.8.8.
+* use predefined compiler version variable, if available.
+
 ## 1.0.2 (2012-03-16)
 
-* Fix Xcode 4.3 compatibility.
+* fix Xcode 4.3 compatibility.
 * Cotire manual corrections.
 
 ## 1.0.1 (2012-03-15)